requirements.txt
diffusers==0.32.2
easydict==1.13
nltk==3.9.1
numpy==2.3.5
omegaconf==2.3.0
pandas==2.3.3
Pillow==12.0.0
pyiqa==0.1.13
pytorch_lightning==2.4.0
PyYAML==6.0.3
timm==1.0.22
tools==1.0.15
torchmetrics==1.4.3
tqdm==4.67.1


